@@ -1,76 +1,125 @@
 {
   "name": "pixiv.moe",
-  "version": "0.3.0",
+  "version": "1.4.0",
   "description": "Source code of https://pixiv.moe",
-  "main": "",
   "scripts": {
     "clean": "rimraf dist/*",
-    "copy": "copyfiles -f ./src/index.html ./src/favicon.ico ./src/robots.txt ./dist",
-    "dist": "npm run copy & webpack --env=dist",
-    "lint": "eslint ./src",
+    "predist": "npm run clean",
+    "dist": "cross-env NODE_ENV=production webpack --config build/webpack.config.js",
+    "postdist": "copyfiles -f ./src/icons/*.ico ./src/icons/*.png ./dist/assets && copyfiles -f ./src/manifest.json ./dist",
+    "lint": "eslint --max-warnings 0 ./src",
+    "lint:fix": "eslint --fix ./src",
     "posttest": "npm run lint",
-    "release:major": "npm version major && npm publish && git push --follow-tags",
-    "release:minor": "npm version minor && npm publish && git push --follow-tags",
-    "release:patch": "npm version patch && npm publish && git push --follow-tags",
-    "serve": "node server.js --env=dev",
-    "serve:dist": "node server.js --env=dist",
-    "start": "node server.js --env=dev",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "cross-env NODE_ENV=development webpack-dev-server --progress --config build/webpack.config.js",
+    "test": "cross-env NODE_ENV=test karma start",
+    "test:watch": "cross-env NODE_ENV=test karma start --autoWatch=true --singleRun=false",
+    "deploy": "gh-pages -d dist",
+    "coverage": "cat ./coverage/lcov.info | coveralls",
+    "contributors:add": "all-contributors add",
+    "contributors:generate": "all-contributors generate"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/kokororin/pixiv.moe.git"
+    "url": "https://github.com/LoveLiveSunshine/pixiv.moe.git"
   },
-  "keywords": [],
+  "keywords": [
+    "pixiv",
+    "lovelive"
+  ],
   "author": "kokororin",
   "license": "MIT",
-  "homepage": "https://github.com/kokororin/pixiv.moe#readme",
+  "homepage": "https://github.com/LoveLiveSunshine/pixiv.moe#readme",
   "bugs": {
-    "url": "https://github.com/kokororin/pixiv.moe/issues"
+    "url": "https://github.com/LoveLiveSunshine/pixiv.moe/issues"
   },
   "devDependencies": {
-    "babel-core": "^6.0.0",
-    "babel-eslint": "^6.0.0",
-    "babel-loader": "^6.0.0",
-    "babel-polyfill": "^6.3.14",
-    "babel-preset-es2015": "^6.0.15",
+    "all-contributors-cli": "^5.2.1",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.2.3",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-minify-dead-code-elimination": "^0.4.3",
+    "babel-plugin-minify-guarded-expressions": "^0.4.3",
+    "babel-plugin-transform-async-to-generator": "^6.16.0",
+    "babel-plugin-transform-decorators-legacy": "^1.3.5",
+    "babel-plugin-transform-object-assign": "^6.8.0",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.8",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.6.0",
     "babel-preset-react": "^6.0.15",
     "babel-preset-stage-0": "^6.5.0",
-    "bower-webpack-plugin": "^0.1.9",
-    "chai": "^3.2.0",
+    "chai": "^4.1.1",
+    "chai-as-promised": "^7.1.1",
     "copyfiles": "^1.0.0",
-    "css-loader": "^0.23.0",
-    "eslint": "^3.0.0",
-    "eslint-loader": "^1.0.0",
-    "eslint-plugin-react": "^6.0.0",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "glob": "^7.0.0",
-    "isparta-instrumenter-loader": "^1.0.0",
+    "coveralls": "^2.13.1",
+    "cross-env": "^5.2.0",
+    "eslint": "^4.5.0",
+    "eslint-config-kotori": "^0.2.3",
+    "eslint-loader": "^1.9.0",
+    "eslint-plugin-prettier": "^2.6.1",
+    "file-loader": "^0.11.2",
+    "gh-pages": "^2.0.1",
+    "glob": "^7.1.2",
+    "html-webpack-plugin": "^3.2.0",
+    "imports-loader": "^0.7.1",
+    "istanbul-instrumenter-loader": "^3.0.0",
+    "karma": "^1.7.0",
+    "karma-chai": "^0.1.0",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-coverage-istanbul-reporter": "^1.3.0",
+    "karma-mocha": "^1.0.0",
+    "karma-mocha-reporter": "^2.0.0",
+    "karma-sourcemap-loader": "^0.3.5",
+    "karma-webpack": "^2.0.4",
     "minimist": "^1.2.0",
-    "mocha": "^3.0.0",
+    "mocha": "^3.5.0",
+    "nock": "^9.0.14",
     "null-loader": "^0.1.1",
     "open": "0.0.5",
-    "phantomjs-prebuilt": "^2.0.0",
-    "react-addons-test-utils": "^15.0.0",
-    "react-hot-loader": "^1.2.9",
+    "open-browser-webpack-plugin": "^0.0.5",
+    "optimize-js-plugin": "^0.0.4",
+    "prettier": "^1.13.6",
+    "puppeteer": "^1.5.0",
+    "react-hot-loader": "next",
+    "react-test-renderer": "^16.4.1",
     "rimraf": "^2.4.3",
-    "style-loader": "^0.13.0",
-    "url-loader": "^0.5.6",
-    "webpack": "^1.12.0",
-    "webpack-dev-server": "^1.12.0"
+    "sw-precache-webpack-plugin": "^0.11.5",
+    "url-loader": "^0.5.9",
+    "webpack": "^3.5.5",
+    "webpack-bundle-analyzer": "^2.13.1",
+    "webpack-dev-server": "^2.7.1",
+    "webpack-manifest-plugin": "^2.0.3",
+    "webpack-merge": "^4.1.3"
   },
   "dependencies": {
-    "core-js": "^2.0.0",
-    "fetch-jsonp": "^1.0.1",
-    "history": "^3.2.1",
-    "imports-loader": "^0.6.5",
-    "react": "^15.0.0",
-    "react-dom": "^15.0.0",
-    "react-ga": "^2.1.1",
-    "react-image-lightbox": "^3.1.0",
-    "react-masonry-component": "^4.2.1",
-    "react-router": "^2.7.0"
+    "@material-ui/core": "^1.3.0",
+    "@material-ui/docs": "^1.0.0-alpha.3",
+    "@material-ui/icons": "^1.1.0",
+    "animation-frame": "^0.3.0",
+    "autobind-decorator": "^1.4.1",
+    "classnames": "^2.2.6",
+    "honoka": "^0.4.17",
+    "inobounce": "^0.1.6",
+    "jss-vendor-prefixer": "^8.0.0",
+    "jszip": "^3.1.4",
+    "moment": "^2.16.0",
+    "namespaced-types": "^0.1.2",
+    "prop-types": "^15.6.2",
+    "react": "^16.4.1",
+    "react-document-title": "^2.0.3",
+    "react-dom": "^16.4.1",
+    "react-event-listener": "^0.6.1",
+    "react-ga": "^2.1.2",
+    "react-image": "^1.0.1",
+    "react-intl": "^2.4.0",
+    "react-masonry-component": "^6.2.1",
+    "react-md-spinner": "^0.2.5",
+    "react-modal": "^3.4.5",
+    "react-redux": "^5.0.6",
+    "react-router-dom": "^4.3.1",
+    "redux": "^4.0.0",
+    "redux-thunk": "^2.3.0",
+    "shortid": "^2.2.6",
+    "store": "^2.0.12"
   },
   "engines": {
     "node": ">=4.0.0"
